SKIP cgroup e2e tests
