+++ tests_e2e/tests/scripts/agent_memory_quota-check_agent_memory_quota.py
@@ -16,40 +16,27 @@
 # limitations under the License.
 #
 import datetime
-import os
 import re
 import sys
 
 from assertpy import fail
 
 from azurelinuxagent.common.future import UTC
-from azurelinuxagent.common.osutil import systemd
 from azurelinuxagent.common.utils import shellutil
 from tests_e2e.tests.lib.agent_log import AgentLog
 from tests_e2e.tests.lib.cgroup_helpers import check_log_message, get_agent_memory_quota, using_cgroupv2, \
-    verify_controllers_available
+    skip_if_memory_controller_is_not_enabled, cleanup_cgroups_test_setup
 
 from tests_e2e.tests.lib.logging import log
 from tests_e2e.tests.lib.remote_test import run_remote_test
 from tests_e2e.tests.lib.retry import retry_if_false
-from tests_e2e.tests.lib.test_result import TestSkipped
 
 
 def skip_if_distro_not_supports_memory_quota():
     if not using_cgroupv2():
         log.info("Skipping  memory quota test as the distro is not using cgroupv2")
-        cleanup_test_setup()
+        cleanup_cgroups_test_setup()
         sys.exit(0)
-
-
-def skip_if_memory_controller_is_not_enabled():
-    # Memory controller only used for report memory metrics. If the controller is not enabled, its ok to skip the test.
-    found: bool = retry_if_false(lambda: verify_controllers_available(["memory"]), delay=120)
-    if not found:
-        cleanup_test_setup()
-        raise TestSkipped("The distro does not have Memory controller enabled. Skipping the test.")
-
-    log.info("Verified memory controller mounted on the system")
 
 
 def prepare_agent():
@@ -156,22 +143,6 @@ def verify_memory_throttling_check_on_agent_cgroups():
     log.info("Successfully verified agent reported zero memory throttling metrics")
 
 
-def cleanup_test_setup():
-    log.info("Cleaning up test setup")
-    drop_in_file = os.path.join(systemd.get_agent_drop_in_path(), "99-ExecStart.conf")
-    if os.path.exists(drop_in_file):
-        log.info("Removing %s...", drop_in_file)
-        os.remove(drop_in_file)
-        shellutil.run_command(["systemctl", "daemon-reload"])
-
-    check_time = datetime.datetime.now(UTC)
-    shellutil.run_command(["agent-service", "restart"])
-
-    found: bool = retry_if_false(lambda: check_log_message(" Agent cgroups enabled: True", after_timestamp=check_time))
-    if not found:
-        fail("Agent cgroups not enabled yet")
-
-
 def main():
     skip_if_distro_not_supports_memory_quota()
     skip_if_memory_controller_is_not_enabled()
@@ -179,7 +150,7 @@ def main():
     verify_agent_has_no_memory_quota_set()
     verify_agent_reported_memory_metrics()
     verify_memory_throttling_check_on_agent_cgroups()
-    cleanup_test_setup()
+    cleanup_cgroups_test_setup()
 
 
 run_remote_test(main)
